Volunteer update
Added all the universitie structure to de model. It's a bit much but it's done now.
The volunteer table is still not finished and wont for a wille. It's a lot of stuff.

Also. Was having a lot of problems making the migrations. Be very carefull with them. It's realy easy to pass the wrong data format into a INT(id).