Going for the html
Fineshed most of the work with the data base. Now I wil work on the Jinja and Html.

Git is teling me that a lot of binary files have changed. Not sure why.

Changes to the header.

Better_Names

Changed some names in the project to try and make them more meaningful.

Fixed nav bar

Now the nav bar will show the active page and proper title.

Fixed the img not showing.

Took me some 2 hours to figure out I was missing a "/" before static. :anguished:

Blog form created.

Added a form or blog. It has a auto_now atribute in the model now so that the time stamp to it will always be the moment it was created.
Did the forms for the DB

It's working but we have problems. I'm still strugling with the date time thing and the many to many models. Also need to fix the names of the templates.
Fixed dates on forms and templates

It was all about this widget thing. I'm still reading the documentation but it's a strugle to get it. Right now only the projects form has the date time selector fixed but I havent changed the formating.

Also changed the name of the templates to something easyer to work with.
Some html things

Trying to make the nav better.
More html

Made the website more cute but I'm strugling a great deal with jquery.
Minor changes to the header

Just some minor changes. The jquery still doesnot work.